Recursively instrument every GC path to capture all sets appropriately. Improve DOM leak support.
Now if window.foo.bar.baz is the target, and the application overwrites window.foo, everything should work fine.

Prior to snapshotting, we call $$$SERIALIZE_DOM$$$ on the agent living in the webpage, which sets up a mirror of the DOM's tree structure as the global variable $$$DOM$$$. The leak-finding code restricts itself to finding leaks reachable from the global scope, and ignores the snapshot's DOM contents. The agent translates any leak paths stemming from $$$DOM$$$ into an actual path in the webpage's DOM to the target element.

Not yet supported: Re-proxying DOM objects when they are re-assigned.